Fixed PHP duplicate interface method (when used with UserBundle)
